refactor: rollEncounter を encounter.ts に分離、snapStep で prevTile 更新
